master_game_agent/pyproject.toml
[project]
name = "master-game-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
